Add shopping list categorization, meal planner improvements, and image upload
- Updated shopping list to have ingredient categorization (Produce, Proteins, Dairy, etc.)
- Fixed ingredient parsing to prevent spacing issues
- Fixed and updated feature to consolidate similar sizes (now shows 3x 1.5 tspn = 4.5 tspn)
- Made ingredient names bold
- Added Dessert/Snacking meal slot to weekly planner
- Removed tag limit in meal planner (now shows all tags)
- Added randomize button for meal planning based on recipe tags (e.g. Breakfast-tagged recipes for breakfast)
- Added clear all button to meal planner
- Added feature to upload images with base64 encoding

+++ backend/app/models/ingredient.py
@@ -1,4 +1,4 @@
-from sqlalchemy import String, BigInteger
+from sqlalchemy import String, Integer
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.db.base import Base
 
@@ -6,7 +6,7 @@ from app.db.base import Base
 class Ingredient(Base):
     __tablename__ = "ingredients"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     name: Mapped[str] = mapped_column(String(255), unique=True, nullable=False)
 
     # Relationships
