+++ backend/app/models/recipe_ingredient.py
@@ -1,4 +1,4 @@
-from sqlalchemy import String, BigInteger, ForeignKey, Numeric, Index
+from sqlalchemy import String, Integer, ForeignKey, Numeric, Index
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.db.base import Base
 
@@ -10,10 +10,10 @@ class RecipeIngredient(Base):
     )
 
     recipe_id: Mapped[int] = mapped_column(
-        BigInteger, ForeignKey("recipes.id", ondelete="CASCADE"), primary_key=True
+        Integer, ForeignKey("recipes.id", ondelete="CASCADE"), primary_key=True
     )
     ingredient_id: Mapped[int] = mapped_column(
-        BigInteger, ForeignKey("ingredients.id", ondelete="CASCADE"), primary_key=True
+        Integer, ForeignKey("ingredients.id", ondelete="CASCADE"), primary_key=True
     )
     quantity: Mapped[float | None] = mapped_column(Numeric(10, 2), nullable=True)
     unit: Mapped[str | None] = mapped_column(String(50), nullable=True)
