+++ backend/app/models/shopping_list.py
@@ -1,4 +1,4 @@
-from sqlalchemy import BigInteger, ForeignKey, Date, DateTime, UniqueConstraint, func
+from sqlalchemy import Integer, ForeignKey, Date, DateTime, UniqueConstraint, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from datetime import date, datetime
 from app.db.base import Base
@@ -10,8 +10,8 @@ class ShoppingList(Base):
         UniqueConstraint("user_id", "week_start_date", name="uq_shopping_lists_user_id_week_start_date"),
     )
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    user_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("users.id"), nullable=False)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    user_id: Mapped[int] = mapped_column(Integer, ForeignKey("users.id"), nullable=False)
     week_start_date: Mapped[date] = mapped_column(Date, nullable=False)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now(), nullable=False)
